Move Compose popup sizing into adapter
Derive the pre-render measure from the synthetic Compose root's wrap-content layout params instead of checking the rendering mode in PaparazziSdk.

Keep the popup-root fallback behavior in ComposeViewAdapter so the raw snapshot loop no longer owns the overlay sizing details.
